backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.4.2
python-jose==3.3.0
groq==0.4.2
python-multipart==0.0.6
logging==0.4.9.6
pydantic-settings==2.0.3
websockets==12.0
typing-extensions==4.9.0
jinja2==3.1.2
aiofiles==23.2.1
